address review: extract shared getInstanceId to instances.js
DRY: move getInstanceId() from brainStorage.js and memoryDB.js to
instances.js where getSelf() already lives, eliminating duplication
